Review and refactoring
