+++ src/common/CMakeLists.txt
@@ -32,7 +32,6 @@ set(efu_common_SRC
 set(efu_common_INC
   ${VERSION_INCLUDE_DIR}/common/version_num.h
   ${VERSION_INCLUDE_DIR}/common/Version.h
-  dataflow/DataObserverTemplate.h
   debug/Assert.h
   debug/Expect.h
   debug/Hexdump.h
