+++ src/modules/timepix3/CMakeLists.txt
@@ -10,6 +10,7 @@ get_filename_component(TIMEPIX_CONFIG "${ESS_MODULE_DIR}/timepix3/configs/timepi
 set(timepix3_common_inc
   Timepix3Base.h
   geometry/Config.h
+  dataflow/DataObserverTemplate.h
   geometry/Timepix3Geometry.h
   readout/DataEventTypes.h
   readout/TimingEventHandler.h
@@ -19,7 +20,6 @@ set(timepix3_common_inc
 set(timepix3_common_src
   Timepix3Base.cpp
   readout/TimingEventHandler.cpp
-  readout/DataEventTypes.cpp
   readout/DataParser.cpp
   Timepix3Instrument.cpp
   geometry/Config.cpp
