--- src/modules/timepix3/readout/DataEventTypes.cpp
@@ -1,15 +1,0 @@
-// Copyright (C) 2023 European Spallation Source, see LICENSE file
-//===----------------------------------------------------------------------===//
-///
-/// \file
-///
-/// \brief Listener interface for data events
-//===----------------------------------------------------------------------===//
-
-#include "DataEventTypes.h"
-#include <cstdint>
-
-
-namespace Timepix3 {
-
-}
