+++ src/modules/timepix3/readout/TimingEventHandler.h
@@ -1,7 +1,15 @@
+// Copyright (C) 2023 European Spallation Source, see LICENSE file
+//===----------------------------------------------------------------------===//
+///
+/// \file
+///
+/// \brief Implementation for TDC and EVR timing event observers. Responsible
+//         to follow up different timing events and syncronize the two timing 
+//         (ESS time and camera time) domains. Also this class provides timing
+//         information for other timing interested objects.
+//===----------------------------------------------------------------------===//
 
-
-
-#include <common/dataflow/DataObserverTemplate.h>
+#include <dataflow/DataObserverTemplate.h>
 #include <readout/DataEventTypes.h>
 #include <vector>
 #include <memory>
