+++ src/modules/timepix3/test/Timepix3ParserTest.cpp
@@ -4,8 +4,6 @@
 /// \file
 //===----------------------------------------------------------------------===//
 
-#include "common/dataflow/DataObserverTemplate.h"
-#include "readout/DataEventTypes.h"
 #include "readout/TimingEventHandler.h"
 #include "gtest/gtest-death-test.h"
 #include "gtest/gtest.h"
@@ -71,14 +69,13 @@ std::vector<uint8_t> TDCAndPixelReadout{
 };
 // clang-format on
 
-
 class Timepix3ParserTest : public TestBase {
 protected:
   struct Counters counters;
   TimingEventHandler testEventHandler;
   DataParser Timepix3Parser{counters, testEventHandler};
 
-  void SetUp() override {counters = {}; }
+  void SetUp() override { counters = {}; }
   void TearDown() override {}
 };
 
@@ -86,14 +83,14 @@ protected:
 
 TEST_F(Timepix3ParserTest, SinglePixelReadout) {
   auto Res = Timepix3Parser.parse((char *)SinglePixelReadout.data(),
-                                   SinglePixelReadout.size());
+                                  SinglePixelReadout.size());
   EXPECT_EQ(Res, 1);
   EXPECT_EQ(counters.PixelReadouts, 1);
 }
 
 TEST_F(Timepix3ParserTest, TDCReadouts) {
   auto Res = Timepix3Parser.parse((char *)TDC1RisingReadout.data(),
-                                   TDC1RisingReadout.size());
+                                  TDC1RisingReadout.size());
   EXPECT_EQ(Res, 1);
   EXPECT_EQ(counters.TDCReadouts, 1);
   EXPECT_EQ(counters.TDC1RisingReadouts, 1);
@@ -108,7 +105,7 @@ TEST_F(Timepix3ParserTest, TDCReadouts) {
   EXPECT_TRUE(testEventHandler.getLastTdcEvent()->Stamp == 6);
 
   Res = Timepix3Parser.parse((char *)TDC1FallingReadout.data(),
-                                   TDC1FallingReadout.size());
+                             TDC1FallingReadout.size());
   EXPECT_EQ(Res, 1);
   EXPECT_EQ(counters.TDCReadouts, 2);
   EXPECT_EQ(counters.TDC1RisingReadouts, 1);
@@ -123,7 +120,7 @@ TEST_F(Timepix3ParserTest, TDCReadouts) {
   EXPECT_TRUE(testEventHandler.getLastTdcEvent()->Stamp == 6);
 
   Res = Timepix3Parser.parse((char *)TDC2RisingReadout.data(),
-                                   TDC2RisingReadout.size());
+                             TDC2RisingReadout.size());
   EXPECT_EQ(Res, 1);
   EXPECT_EQ(counters.TDCReadouts, 3);
   EXPECT_EQ(counters.TDC1RisingReadouts, 1);
@@ -138,7 +135,7 @@ TEST_F(Timepix3ParserTest, TDCReadouts) {
   EXPECT_TRUE(testEventHandler.getLastTdcEvent()->Stamp == 6);
 
   Res = Timepix3Parser.parse((char *)TDC2FallingReadout.data(),
-                                   TDC2FallingReadout.size());
+                             TDC2FallingReadout.size());
   EXPECT_EQ(Res, 1);
   EXPECT_EQ(counters.TDCReadouts, 4);
   EXPECT_EQ(counters.TDC1RisingReadouts, 1);
@@ -160,14 +157,14 @@ TEST_F(Timepix3ParserTest, TooShort) {
 
 TEST_F(Timepix3ParserTest, SingleEVRReadout) {
   auto Res = Timepix3Parser.parse((char *)SingleEVRReadout.data(),
-                                   SingleEVRReadout.size());
+                                  SingleEVRReadout.size());
   EXPECT_EQ(Res, 1);
   EXPECT_EQ(counters.EVRTimestampReadouts, 1);
 }
 
 TEST_F(Timepix3ParserTest, TDCAndPixelReadout) {
   auto Res = Timepix3Parser.parse((char *)TDCAndPixelReadout.data(),
-                                   TDCAndPixelReadout.size());
+                                  TDCAndPixelReadout.size());
   EXPECT_EQ(Res, 2);
   EXPECT_EQ(counters.TDCReadouts, 1);
   EXPECT_EQ(counters.PixelReadouts, 1);
@@ -178,7 +175,6 @@ TEST_F(Timepix3ParserTest, TDCAndPixelReadout) {
   EXPECT_TRUE(testEventHandler.getLastTdcEvent()->TriggerCounter == 2231);
   EXPECT_TRUE(testEventHandler.getLastTdcEvent()->Timestamp == 31447764897);
   EXPECT_TRUE(testEventHandler.getLastTdcEvent()->Stamp == 6);
-  
 }
 
 int main(int argc, char **argv) {
